Changed hash table to be fully allocated on program start or hash clear/size change.

+++ search.cpp
@@ -810,7 +810,7 @@ void clearTables() {
     searchParams.resetHistoryTable();
 }
 
-void setHashSize(int MB) {
+void setHashSize(uint64_t MB) {
     transpositionTable.setSize(MB);
 }
