+++ search.h
@@ -25,7 +25,7 @@
 
 void getBestMove(Board *b, int mode, int value, Move *bestMove);
 void clearTables();
-void setHashSize(int MB);
+void setHashSize(uint64_t MB);
 uint64_t getNodes();
 
 int getBestMoveForSort(Board *b, MoveList &legalMoves, int depth);
